rename: 컴포넌트 폴더 분리 및 파일 이사
