+++ src/app/(public)/login/LoginForm.tsx
@@ -4,7 +4,7 @@ import { useMutation } from "@tanstack/react-query";
 import { useForm } from "react-hook-form";
 import { authApi } from "@/api";
 import { useHandleAuth } from "@/services/auth";
-import { Spinner, customToast } from "@/components";
+import { Spinner, customToast } from "@/components/ui";
 import { PostLoginRequestBody } from "@/types/api";
 
 export default function LoginForm() {
