+++ src/app/layout.tsx
@@ -3,8 +3,7 @@ import { Nanum_Gothic } from "next/font/google";
 import { cookies } from "next/headers";
 import { apiServer } from "@/api/instance";
 import TanStackQueryProvider from "@/providers/TanStackQueryProvider";
-import { AnimationContainer } from "@/components";
-import { CustomToastContainer } from "@/components/CustomToast";
+import { AnimationContainer, CustomToastContainer } from "@/components/ui";
 import "@/styles/globals.css";
 
 export const metadata: Metadata = {
