+++ src/components/form/CharacterSelect.tsx
