+++ src/components/form/index.ts
@@ -1,0 +1,6 @@
+export { default as CharacterSelect } from "./CharacterSelect";
+export { default as CheckBox } from "./CheckBox";
+export { default as CheckList } from "./CheckList";
+export { default as ColorSelect } from "./ColorSelect";
+export { default as FixedInput } from "./FixedInput";
+export { default as LargeInput } from "./LargeInput";
