--- src/components/index.ts
@@ -1,15 +1,0 @@
-export { default as CharacterSelect } from "./CharacterSelect";
-export { default as CheckBox } from "./CheckBox";
-export { default as CheckList } from "./CheckList";
-export { default as ColorSelect } from "./ColorSelect";
-export { default as CustomCharacter } from "./CustomCharacter";
-export { default as LargeInput } from "./LargeInput";
-export { default as LinkButton } from "./LinkButton";
-export { default as Triangle } from "./Triangle";
-export { default as BallonBox } from "./BallonBox";
-export { default as CustomIcon } from "./CustomIcon";
-export { default as ShowError } from "./ShowError";
-export { default as Spinner } from "./Spinner";
-export { default as FixedInput } from "./FixedInput";
-export * from "./CustomToast";
-export * from "./ShowAnimation";
