+++ src/components/ui/CustomCharacter.tsx
