+++ src/components/ui/ShowAnimation.tsx
